chore: move factory methods into class it creates
