+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/report/TrackerErrorReport.java
@@ -119,6 +119,11 @@ public class TrackerErrorReport
         }
     }
 
+    public static TrackerErrorReportBuilder newReport( TrackerErrorCode errorCode )
+    {
+        return builder().errorCode( errorCode );
+    }
+
     @Override
     public String toString()
     {
