+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/report/TrackerWarningReport.java
@@ -116,6 +116,11 @@ public class TrackerWarningReport
         }
     }
 
+    public static TrackerWarningReportBuilder newWarningReport( TrackerErrorCode errorCode )
+    {
+        return builder().warningCode( errorCode );
+    }
+
     @Override
     public String toString()
     {
